refactor: mvc login users
